+++ email.css
@@ -1,0 +1,75 @@
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f2f2f2;
+}
+
+.container {
+    max-width: 400px;
+    margin: auto;
+    padding: 50px;
+    margin-top: 200px;
+    text-align: center;
+}
+
+.logo {
+    position: absolute;
+    top: 15px;
+    left: 15px;
+    width: 120px; 
+    height: auto;
+}
+
+
+h2 {
+    margin-bottom: 20px;
+    color: #1E4F66;
+}
+
+p {
+    margin-bottom: 20px;
+    color: #7C7C7C;
+}
+
+.verification-code {
+    margin-bottom: 20px;
+}
+
+.verification-code input[type="text"] {
+    width: 30px;
+    height: 30px;
+    text-align: center;
+    margin-right: 5px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+.buttons {
+    margin-top: 20px;
+}
+
+.buttons button {
+    padding: 10px 20px;
+    margin-right: 10px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+.resend {
+    font-size: 20px;
+    background-color: #7F9DAB;
+    color: white;
+}
+.resend:hover { 
+    background-color: #0056b3;
+}
+
+
+.verify {
+    font-size: 20px;
+    background-color: #7F9DAB;;
+    color: white;
+}
+.verify:hover { 
+    background-color: #0056b3;
+}
