+++ email.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Email Verification</title>
+    <link rel="stylesheet" href="email.css">
+</head>
+<body>
+    <img src="logo.png" alt="Logo" class="logo">
+    <div class="container">
+        <h2>Verify Your Email</h2>
+        <p>We’ve sent a verification 6 digits code to your email address. 
+            Please enter the code to complete the email verification process.</p>
+        <form action="#">
+            <div class="verification-code">
+                <input type="text" maxlength="1" required>
+                <input type="text" maxlength="1" required>
+                <input type="text" maxlength="1" required>
+                <input type="text" maxlength="1" required>
+                <input type="text" maxlength="1" required>
+                <input type="text" maxlength="1" required>
+            </div>
+            <div class="buttons">
+                <button type="button" class="resend">Resend Code</button>
+                <button type="submit" class="verify">Verify</button>
+            </div>
+        </form>
+    </div>
+</body>
+</html>
