+++ new.css
@@ -1,0 +1,64 @@
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f2f2f2;
+}
+
+.container {
+    max-width: 400px;
+    margin: auto;
+    padding: 50px;
+    margin-top: 200px;
+    /* text-align: center; */
+}
+
+.logo {
+    position: absolute;
+    top: 15px;
+    left: 15px;
+    width: 120px; 
+    height: auto;
+}
+
+h2 {
+    margin-bottom: 20px;
+    color: #1E4F66;
+    margin-top: 30px;
+    margin-bottom: 10px;
+    text-align: center;
+}
+
+p {
+    margin-bottom: 20px;
+    color: #7C7C7C;
+    text-align: center;
+}
+
+label {
+    display: block;
+    margin-bottom: 10px;
+}
+
+input[type="email"],
+input[type="password"] {
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 15px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    box-sizing: border-box;
+}
+
+input[type="submit"] {
+    font-size: 20px;
+    width: 100%;
+    background-color: #7F9DAB;
+    color: white;
+    padding: 10px 0;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+input[type="submit"]:hover {
+    background-color: #0056b3;
+}
